fix : flyway 및 friend, friendrequest 합침
