+++ backend/src/main/kotlin/com/puzzle/backend/common/config/SecurityConfig.kt
@@ -30,12 +30,13 @@ class SecurityConfig(
             .sessionManagement { it.sessionCreationPolicy(SessionCreationPolicy.STATELESS) } // 세션을 사용하지 않도록 설정 (JWT로 인증하기 때문)
             .authorizeHttpRequests {
                 // 요청 URL별 권한 설정
-                it
-                    .requestMatchers(
-                        *PERMITTED_URL_PATTERNS,
-                    ).permitAll()
-                    .anyRequest()
-                    .authenticated() // 나머지 모든 요청은 인증 확인
+//                it
+//                    .requestMatchers(
+//                        *PERMITTED_URL_PATTERNS,
+//                    ).permitAll()
+//                    .anyRequest()
+//                    .authenticated() // 나머지 모든 요청은 인증 확인
+                it.anyRequest().permitAll()
             }.oauth2Login {
                 it.userInfoEndpoint { it.userService(customOAuth2UserService::loadUser) }
                 it.successHandler(oAuth2AuthenticationSuccessHandler)
