+++ backend/src/main/kotlin/com/puzzle/backend/friend/domain/Friend.kt
@@ -1,7 +1,10 @@
 package com.puzzle.backend.friend.domain
 
+import com.puzzle.backend.common.domain.BaseEntity
 import com.puzzle.backend.oauth.domain.Users
 import jakarta.persistence.Entity
+import jakarta.persistence.EnumType
+import jakarta.persistence.Enumerated
 import jakarta.persistence.GeneratedValue
 import jakarta.persistence.GenerationType
 import jakarta.persistence.Id
@@ -15,10 +18,20 @@ class Friend(
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     var id: Long = 0,
+    // 요청을 보낸 사용자
     @ManyToOne
-    @JoinColumn(name = "user_id")
-    var user: Users,
+    @JoinColumn(name = "requester_id")
+    var requester: Users,
+    // 요청을 받은 사용자
     @ManyToOne
-    @JoinColumn(name = "friend_id")
-    var friend: Users,
-)
+    @JoinColumn(name = "receiver_id")
+    var receiver: Users,
+    // 요청의 상태 (PENDING, ACCEPTED)
+    @Enumerated(EnumType.STRING)
+    var status: RequestStatus = RequestStatus.PENDING,
+) : BaseEntity()
+
+enum class RequestStatus {
+    PENDING,
+    ACCEPTED,
+}
